Fix empty file crashes

+++ utils/src/main/java/com/omar/retromp3recorder/utils/Mp3TagsEditor.kt
@@ -22,23 +22,25 @@ class Mp3TagsEditorImpl @Inject constructor(
 ) : Mp3TagsEditor {
     override fun setTags(filepath: String, tags: RecordingTags) {
         if (isTagsWork().not()) return
-        val mp3File = Mp3File(filepath)
-        mp3File.id3v1Tag = ID3v1Tag().apply {
-            year = tags.year
-            artist = tags.artist
-            title = tags.title
+        if (fileEmptyChecker.isFileEmpty(filepath).not()) {
+            val mp3File = Mp3File(filepath)
+            mp3File.id3v1Tag = ID3v1Tag().apply {
+                year = tags.year
+                artist = tags.artist
+                title = tags.title
+            }
+            val temp = "${context.cacheDir}/temp.mp3"
+            mp3File.save(temp)
+            try {
+                File(temp).copyTo(File(filepath), overwrite = true)
+                File(temp).delete()
+            } catch (e: FileAlreadyExistsException) {
+                Timber.e(e)
+            }
+            val newFile = Mp3File(filepath)
+            val id3v1Tag = newFile.id3v1Tag
+            Timber.d("Tag $id3v1Tag")
         }
-        val temp = "${context.cacheDir}/temp.mp3"
-        mp3File.save(temp)
-        try {
-            File(temp).copyTo(File(filepath), overwrite = true)
-            File(temp).delete()
-        } catch (e: FileAlreadyExistsException) {
-            Timber.e(e)
-        }
-        val newFile = Mp3File(filepath)
-        val id3v1Tag = newFile.id3v1Tag
-        Timber.d("Tag $id3v1Tag")
     }
 
     override fun getFilenameFromPath(filePath: String): String {
